tests: dirty fix for cellToPolygon tests
Since the internal implementation of the cellToPolygon algorithm now
uses an HashMap instead of a BTreeMap, the polygons' loops may change
between two executions (they are the same loop geometrically speaking,
but the point chosen as the start point may differ).

Since the test was expecting a given order, it's now broken/flaky.
Until we find a better solution, we sort the output to have a green test.

This ensure that we have at least the right coordinates in the output.
But we've lost checks on:
- the lat/lng pairing
- the relative order inside a ring
- the rings are closed